refactor(errors,cli,a2h): collapse RemoteRPCError twin, split cli, fix a2h ordering (S3)
Sprint S3 Task 6.0 — models+misc cleanup:

- Collapse RemoteFatalRPCError/RemoteRecoverableRPCError into one
  RemoteRPCError(ASAPError) with an is_recoverable property
  (__init__/__str__/data/from_jsonrpc/asap_slot implemented once; ~55 LOC
  deleted). The twin names are kept as deprecated subclasses
  (RemoteFatalRPCError is_recoverable=False; RemoteRecoverableRPCError
  is_recoverable=True) so chaos-suite isinstance/exception-tuples stay
  green with zero churn. transport/errors.py docstring repointed to
  RemoteRPCError (was referencing the removed twin "until the roadmap
  collapse").
- Split cli/__init__.py (596 LOC) into per-command-group modules
  (schemas/keys/manifest/delegation/trace/repl) following the existing
  register_* pattern; __init__.py becomes Typer wiring + main(). Legacy
  public symbols (DEFAULT_SCHEMAS_DIR, export_all_schemas, _repl_namespace)
  re-exported so tests/test_cli.py keeps resolving. Annotated[...] params
  standardized across commands.
- Fix a2h.py RenderContent forward-ref ordering (move RenderContent +
  ChannelFallback above ChannelBinding so the render field is no longer
  a forward ref).

Behavior-preserving; public surface stable (deprecated subclass aliases +
cli legacy re-exports). The _send.py retry-decision rewrite
(isinstance -> is_recoverable) lands in the transport commit alongside
the codecs inline.
